Fix #3600
Former-commit-id: f26adc053068e3442c2613cc645d724d7388c255
Former-commit-id: 5841108f16f2d4aa4e17018dc0907d45b29a3d46

+++ web/concrete/src/Page/Type/Type.php
@@ -218,17 +218,19 @@ class Type extends Object implements \Concrete\Core\Permission\ObjectInterface
         }
 
         // remove all but the most recent X drafts.
-        $vl = new VersionList($c);
-        $vl->setItemsPerPage(-1);
-        // this will ensure that we only ever keep X versions.
-        $vArray = $vl->getPage();
-        if (count($vArray) > $this->ptDraftVersionsToSave) {
-            for ($i = $this->ptDraftVersionsToSave; $i < count($vArray); $i++) {
-                $v = $vArray[$i];
-                @$v->delete();
+        if ($c->isPageDraft()) {
+            $vl = new VersionList($c);
+            $vl->setItemsPerPage(-1);
+            // this will ensure that we only ever keep X versions.
+            $vArray = $vl->getPage();
+            if (count($vArray) > $this->ptDraftVersionsToSave) {
+                for ($i = $this->ptDraftVersionsToSave; $i < count($vArray); $i++) {
+                    $v = $vArray[$i];
+                    @$v->delete();
+                }
             }
         }
-
+        
         $c = Page::getByID($c->getCollectionID(), 'RECENT');
         $controls = array();
         foreach ($outputControls as $oc) {
